Changes after review.

+++ src/main/java/com/pubnub/api/managers/DelayedReconnectionManager.java
@@ -64,7 +64,7 @@ public class DelayedReconnectionManager {
         return false;
     }
 
-    private double getRandomDelayInMilliSeconds() {
+    private int getRandomDelayInMilliSeconds() {
        return random.nextInt(BOUND);
     }
